+++ examples/spark/src/test/java/edu/snu/nemo/examples/spark/MRJava.java
@@ -13,14 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package edu.snu.nemo.examples.spark;
 
 import edu.snu.nemo.client.JobLauncher;
 import edu.snu.nemo.common.test.ArgBuilder;
 import edu.snu.nemo.common.test.ExampleTestUtil;
 import edu.snu.nemo.compiler.optimizer.policy.DefaultPolicy;
-import edu.snu.nemo.examples.spark.sql.JavaUserDefinedTypedAggregation;
-import edu.snu.nemo.examples.spark.sql.JavaUserDefinedUntypedAggregation;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -29,12 +28,12 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
 /**
- * Test Spark programs with JobLauncher.
+ * Test MR Spark programs with JobLauncher.
  */
 @RunWith(PowerMockRunner.class)
 @PrepareForTest(JobLauncher.class)
 @PowerMockIgnore("javax.management.*")
-public final class SparkJavaITCase {
+public final class MRJava {
   private static final int TIMEOUT = 180000;
   private static ArgBuilder builder;
   private static final String fileBasePath = System.getProperty("user.dir") + "/../resources/";
@@ -71,9 +70,9 @@ public final class SparkJavaITCase {
   /* Temporary disabled because of Travis issue
   @Test(timeout = TIMEOUT)
   public void testSparkMapReduce() throws Exception {
-    final String inputFileName = "sample_input_mr";
+    final String inputFileName = "sample_input_wordcount_spark";
     final String outputFileName = "sample_output_mr";
-    final String testResourceFileName = "test_output_mr";
+    final String testResourceFilename = "test_output_wordcount_spark";
     final String inputFilePath = fileBasePath + inputFileName;
     final String outputFilePath = fileBasePath + outputFileName;
     final String parallelism = "2";
@@ -87,63 +86,9 @@ public final class SparkJavaITCase {
         .build());
 
     try {
-      ExampleTestUtil.ensureOutputValidity(fileBasePath, outputFileName, testResourceFileName);
+      ExampleTestUtil.ensureOutputValidity(fileBasePath, outputFileName, testResourceFilename);
     } finally {
       ExampleTestUtil.deleteOutputFile(fileBasePath, outputFileName);
     }
   }*/
-
-  @Test(timeout = TIMEOUT)
-  public void testSparkPi() throws Exception {
-    final String numParallelism = "3";
-
-    JobLauncher.main(builder
-        .addJobId(JavaSparkPi.class.getSimpleName() + "_test")
-        .addUserMain(JavaSparkPi.class.getCanonicalName())
-        .addUserArgs(numParallelism)
-        .addOptimizationPolicy(DefaultPolicy.class.getCanonicalName())
-        .build());
-  }
-
-  @Test(timeout = TIMEOUT)
-  public void testSparkSQLUserDefinedTypedAggregation() throws Exception {
-    final String inputFileName = "sample_input_employees.json";
-    final String inputFilePath = fileBasePath + inputFileName;
-
-    JobLauncher.main(builder
-        .addJobId(JavaUserDefinedTypedAggregation.class.getSimpleName() + "_test")
-        .addUserMain(JavaUserDefinedTypedAggregation.class.getCanonicalName())
-        .addUserArgs(inputFilePath)
-        .addOptimizationPolicy(DefaultPolicy.class.getCanonicalName())
-        .build());
-  }
-
-  @Test(timeout = TIMEOUT)
-  public void testSparkSQLUserDefinedUntypedAggregation() throws Exception {
-    final String inputFileName = "sample_input_employees.json";
-    final String inputFilePath = fileBasePath + inputFileName;
-
-    JobLauncher.main(builder
-        .addJobId(JavaUserDefinedUntypedAggregation.class.getSimpleName() + "_test")
-        .addUserMain(JavaUserDefinedUntypedAggregation.class.getCanonicalName())
-        .addUserArgs(inputFilePath)
-        .addOptimizationPolicy(DefaultPolicy.class.getCanonicalName())
-        .build());
-  }
-
-  @Test(timeout = TIMEOUT)
-  public void testSparkSQLExample() throws Exception {
-    final String peopleJson = "sample_input_people.json";
-    final String peopleTxt = "sample_input_people.txt";
-    final String inputFileJson = fileBasePath + peopleJson;
-    final String inputFileTxt = fileBasePath + peopleTxt;
-
-    //    TODO#12: Frontend support for Scala Spark.
-    //    JobLauncher.main(builder
-    //        .addJobId(JavaSparkSQLExample.class.getSimpleName() + "_test")
-    //        .addUserMain(JavaSparkSQLExample.class.getCanonicalName())
-    //        .addUserArgs(inputFileJson, inputFileTxt)
-    //        .addOptimizationPolicy(DefaultPolicy.class.getCanonicalName())
-    //        .build());
-  }
 }
